\name{hfdata}
\docType{data}
\alias{hfdata}
\title{Hedge-fund data}
\description{%
  The matrix \code{hfdata} contains randomized monthly returns of hedge funds.
}
\usage{data(hfdata)}
\format{A matrix of size 60 x 100, containing 60 monthly net returns for 100 hedge funds. Returns have been slightly modified.}
\source{\url{http://www.hedgefundresearch.com}.}
\references{%
  Ardia, D., Boudt, K. (2014).
  \emph{The Peer Performance of Hedge Funds}.
  \url{http://papers.ssrn.com/sol3/papers.cfm?abstract_id=2000901}

  Guidotti, I., Nagui, I. (2012).
  \emph{Rational Behavior of Hedge Fund Managers and Investors}.
  \url{http://papers.ssrn.com/sol3/papers.cfm?abstract_id=1972148}
}
\keyword{datasets}




