% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.exposureHeterogeneity}
\alias{plot.exposureHeterogeneity}
\title{Plot factor exposure heterogeneity}
\usage{
\method{plot}{exposureHeterogeneity}(x, ...)
}
\arguments{
\item{x}{An \code{exposureHeterogeneity} object.}

\item{...}{Further arguments passed to \code{\link[graphics]{barplot}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Bar plot of the heterogeneity measure \eqn{1-\pi^0_k} returned by
\code{\link{exposureHeterogeneity}}.
}
\seealso{
\code{\link{exposureHeterogeneity}}.
}
\author{
David Ardia and Kris Boudt.
}
\keyword{hplot}
