% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollScreening.R
\name{plot.rollScreening}
\alias{plot.rollScreening}
\title{Plot a rolling peer performance screening}
\usage{
\method{plot}{rollScreening}(x, what = NULL, legend = TRUE, ...)
}
\arguments{
\item{x}{A \code{rollScreening} object.}

\item{what}{Either \code{"heterogeneity"} (\eqn{1-\bar\pi^0}) or
\code{"ratios"} (the averaged \eqn{\bar\pi^+,\bar\pi^0,\bar\pi^-}). Default:
\code{"heterogeneity"} for a beta screening, \code{"ratios"} otherwise.}

\item{legend}{A logical value indicating whether to draw a legend. Default:
\code{TRUE}.}

\item{...}{Further graphical arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Time-series plot of the rolling peer performance ratios produced
by \code{\link{rollScreening}}.
}
\examples{
\donttest{
  data("hfdata")
  roll <- rollScreening(hfdata[, 1:15], width = 36, by = 12,
                        control = list(nCore = 1))
  plot(roll)
}
}
\seealso{
\code{\link{rollScreening}}.
}
\author{
David Ardia and Kris Boudt.
}
\keyword{hplot}
