% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScreening.R
\name{plot.SCREENING}
\alias{plot.SCREENING}
\title{Plot SCREENING objects}
\usage{
\method{plot}{SCREENING}(x, ...)
}
\arguments{
\item{x}{A SCREENING object (returned by alphaScreening/sharpeScreening/msharpeScreening).}

\item{...}{Passed to \code{plot.pp_screening()}.}
}
\description{
S3 method for \code{plot()} for objects of class \code{"SCREENING"}.
}
