% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.SCREENING}
\alias{print.SCREENING}
\title{Print method for the 'SCREENING' object}
\usage{
\method{print}{SCREENING}(x, ...)
}
\arguments{
\item{x}{A \code{SCREENING} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for the \code{SCREENING} object returned by
\code{\link{alphaScreening}}, \code{\link{sharpeScreening}} and
\code{\link{msharpeScreening}}.
}
\seealso{
\code{\link{alphaScreening}}, \code{\link{sharpeScreening}},
\code{\link{msharpeScreening}} and \code{\link{plot.SCREENING}}.
}
\author{
David Ardia and Kris Boudt.
}
\keyword{htest}
