% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.SCREENING}
\alias{print.summary.SCREENING}
\title{Print method for the 'summary.SCREENING' object}
\usage{
\method{print}{summary.SCREENING}(x, ...)
}
\arguments{
\item{x}{A \code{summary.SCREENING} object.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for the object returned by
\code{\link{summary.SCREENING}}.
}
\author{
Murilo Pedro, David Ardia and Kris Boudt.
}
\keyword{htest}
