% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.TESTING}
\alias{print.TESTING}
\title{Print method for the 'TESTING' object}
\usage{
\method{print}{TESTING}(x, ...)
}
\arguments{
\item{x}{A \code{TESTING} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for the \code{TESTING} object returned by
\code{\link{sharpeTesting}}, \code{\link{msharpeTesting}} and
\code{\link{alphaTesting}}.
}
\seealso{
\code{\link{sharpeTesting}}, \code{\link{msharpeTesting}} and
\code{\link{alphaTesting}}.
}
\author{
David Ardia and Kris Boudt.
}
\keyword{htest}
