% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryScreening.R
\name{summary.SCREENING}
\alias{summary.SCREENING}
\title{Summarize SCREENING objects}
\usage{
\method{summary}{SCREENING}(x, coef = 1L, top = 5L, p_level = 0.05)
}
\arguments{
\item{coef}{If the selected measure is matrix/array-like, which row/coef to use (default 1).}

\item{top}{How many best/worst funds to show.}

\item{p_level}{Significance level for win/loss counts (requires pval+tstat in object).}

\item{object}{A SCREENING object.}
}
\value{
An object of class \code{"summary.SCREENING"}.
}
\description{
S3 method for \code{summary()} for objects of class \code{"SCREENING"}.
}
