% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfdata.R
\docType{data}
\name{hfdata}
\alias{hfdata}
\title{Hedge fund data}
\format{
A matrix of size 60 x 100, containing 60 monthly net returns for 100 hedge funds. Returns have been modified.
}
\source{
HFR database
}
\usage{
data(hfdata)
}
\description{
The matrix \code{hfdata} contains randomized and modified monthly returns of hedge funds.
}
\keyword{datasets}
