% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.exposureHeterogeneity}
\alias{plot.exposureHeterogeneity}
\title{Plot factor exposure heterogeneity}
\usage{
\method{plot}{exposureHeterogeneity}(x, ...)
}
\arguments{
\item{x}{An \code{exposureHeterogeneity} object.}

\item{...}{Further arguments passed to \code{\link[graphics]{barplot}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Bar plot of the heterogeneity measure \eqn{1-\pi^0_k} returned by
\code{\link{exposureHeterogeneity}}.
}
\examples{
\donttest{
  data("hfdata")
  sc <- alphaScreening(hfdata[, 1:10], factors = hfdata[, 50:51],
                       control = list(nCore = 1, screen_beta = TRUE))
  plot(exposureHeterogeneity(sc))
}
}
\seealso{
\code{\link{exposureHeterogeneity}}.
}
\author{
David Ardia and Kris Boudt.
}
\keyword{hplot}
